remove the useless OptixMesh and rename the files
